crawler/requirements.txt
sqlalchemy
kafka
beautifulsoup4==4.8.0
certifi==2019.9.11
chardet==3.0.4
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.8
jieba3k==0.35.1
kafka-python==1.4.6
lxml==4.4.1
newspaper3k==0.2.8
nltk==3.4.5
Pillow==6.1.0
python-dateutil==2.8.0
PyYAML==5.1.2
redis==3.3.8
requests==2.22.0
requests-file==1.4.3
selenium==3.141.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.3
tinysegmenter==0.3
tldextract==2.2.1
urllib3==1.25.3
